correction

+++ main.h
@@ -25,7 +25,7 @@ int print_char(va_list c);
 int print_37(void);
 int print_str(va_list lot);
 int print_hex(va_list lot);
-int print_hex_again(unsigned int num);
+int print_hex_again(unsigned long int num);
 int print_bin(va_list lot);
 int print_custom_string(va_list lot);
 int print_unsigned_integer(va_list lot);
